First beta release of documentation, mostly environment setup
Most of the development tools have been indentified and
documented, and there is a little bit of source/test code for
a Transaction class as well as for some custom type checking.

After figuring out how the fundamental tools work, a bunch of
background/setup tasks were documented so that development
could be performed in a repeatable environment. A sequence
of planned development tasks and associated releases is
planned, with several beta versions culminating in a
minimum viable prototype (MVP) release

Most of this version is a demonstration of Sphinx usage,
and importantly, the identification of tutorial-based
documentation. After "getting in the right jungle", the
stage is now set for "cutting down some trees" (from
Stephen R. Covey, in "The 7 Habits of Highly Effective People")